made app working in previous state
