+++ src/app/home-module/advertisment-details/advertisment-details.component.html
@@ -42,12 +42,10 @@
             <!-- <button class="right-corner"><i class="fa fa-share"></i></button>
                     <button class="right-corner"><i class="fa fa-heart"></i></button> -->
           </span>
-
         </div>
         <div class="other-info">
           <span class="product-title">{{ productdetails.product.name }}</span>
         </div>
-
         <div class="item-add-details" style="border: none;">
           <div>
             <span class="product-location">{{ productdetails.product.pickupAddress.city }}</span>
@@ -58,7 +56,6 @@
             </span>
           </div>
         </div>
-
       </div>
       <div class="seller-details">
         <div class="seller-description">Seller description</div>
@@ -78,8 +75,6 @@
         <!-- <div *ngIf='!isPreviewOn'>
             <button>Chat with Seller</button> 
         </div> -->
-
-
         <div class="seller-contact">
           <!-- <span><i class="fa fa-phone"></i></span>
               <div>** *** ****</div>
@@ -112,6 +107,4 @@
       </div>
     </div>
   </div>
-</div>
-</div>
 </div>
